전체 Total 업데이트 로직 구현, 긴급도&우선도 표기를 실수 형식으로 변경

+++ infra/priority_db.py
@@ -1,6 +1,6 @@
 from datetime import datetime, date
 from typing import List
-from sqlalchemy import Integer, String, DateTime, ForeignKey, Boolean
+from sqlalchemy import Float, Integer, String, DateTime, ForeignKey, Boolean
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column, relationship
 
 class Base(DeclarativeBase):
@@ -38,8 +38,8 @@ class SubTask(Base):
     task_id : Mapped[int] = mapped_column(Integer, ForeignKey("tasks.task_id"))
     subtask_title : Mapped[str] = mapped_column(String(200), nullable=False)
     ratio : Mapped[int] = mapped_column(Integer, nullable=False)
-    urgent : Mapped[int] = mapped_column(Integer, default=3)
-    importance : Mapped[int] = mapped_column(Integer, nullable=False)
+    urgent : Mapped[float] = mapped_column(Float, default=3.0)
+    importance : Mapped[float] = mapped_column(Float, nullable=False)
     order : Mapped[int] = mapped_column(Integer, nullable=False)
     scheduled_date : Mapped[datetime] = mapped_column(DateTime, nullable=True)
     estimated_time : Mapped[int] = mapped_column(Integer, nullable=True)
